Base de datos Alimentos Lista. Busqueda alimentos. Añadir Alimento (Soluccionar error cuando metes un alimento ya existente Cargar datos tabla menu Conexion Comidas OK Guardamos menU Recargar MenBorrar Alimento Borrar Comida Borrar Alimentos Implementar Calendario

+++ src/main/java/com/mycompany/proyecto_ad08/AnadirQuitarComida.java
@@ -4,6 +4,8 @@ import static com.mycompany.proyecto_ad08.BDComidas.con;
 import static com.mycompany.proyecto_ad08.BDComidas.getConnection;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
+import java.sql.Date;
+import javax.swing.JOptionPane;
 
 
 /*
@@ -29,8 +31,8 @@ public class AnadirQuitarComida {
             PreparedStatement pstmt = null;
             pstmt = con.prepareStatement("INSERT INTO ALIMENTOS (FECHA,COMIDA,ALIMENTO,CALORIAS) VALUES (?,?,?,?)");
 
-            pstmt.setObject(1, PanelPrincipal.fechaComida_in.getText());
-            pstmt.setObject(2, PanelPrincipal.comidaComida_in.getText());
+            pstmt.setObject(1, PanelAnadirComida.fechaAnadirComidaIn.getDate());
+            pstmt.setObject(2, PanelAnadirComida.comidaAndirComidaIn.getSelectedItem().toString());
             pstmt.setObject(3, PanelPrincipal.tablaMenu.getValueAt(i, 0));
             pstmt.setObject(4, PanelPrincipal.tablaMenu.getValueAt(i, 1));
             pstmt.execute();
@@ -42,24 +44,27 @@ public class AnadirQuitarComida {
     }
 
     public static void quitar() throws SQLException {
+        
+        int reply = JOptionPane.showConfirmDialog(null, "¿Esta seguro que quiere eliminar la comida seleccionada?", "Confirmar", JOptionPane.YES_NO_OPTION);
+        if (reply == JOptionPane.YES_OPTION) {
 
         con = getConnection();
-
         
          PreparedStatement pstmt = null;
          
         pstmt = con.prepareStatement("DELETE FROM ALIMENTOS WHERE FECHA=? AND COMIDA=?");
 
-        pstmt.setObject(1, (String) BDComidas.modeloComidas.getValueAt(PanelPrincipal.tablaComida.getSelectedRow(), 0));
+        pstmt.setObject(1, (Date) BDComidas.modeloComidas.getValueAt(PanelPrincipal.tablaComida.getSelectedRow(), 0));
         pstmt.setObject(2, (String) BDComidas.modeloComidas.getValueAt(PanelPrincipal.tablaComida.getSelectedRow(), 1));
         pstmt.executeUpdate();
+        
+        JOptionPane.showMessageDialog(null, "La comida seleccionada ha sido borrado");
         
         BDComidas.cargarTabla();
         CrearMenu.nuevoMenu();
 
-        
         con.close();
-        
+        }
 
     }
 }
